Cleanup various warnings encountered when running Satpy tests

+++ satpy/tests/reader_tests/test_fci_l2_nc.py
@@ -75,7 +75,7 @@ class TestFciL2NCFileHandler(unittest.TestCase):
             two_layers_dataset[0, :, :] = np.ones((100, 10))
             two_layers_dataset[1, :, :] = 2 * np.ones((100, 10))
 
-            mtg_geos_projection = nc.createVariable('mtg_geos_projection', np.int, dimensions=())
+            mtg_geos_projection = nc.createVariable('mtg_geos_projection', int, dimensions=())
             mtg_geos_projection.longitude_of_projection_origin = 10.0
             mtg_geos_projection.semi_major_axis = 6378137.
             mtg_geos_projection.semi_minor_axis = 6356752.
@@ -403,7 +403,7 @@ class TestFciL2NCReadingByteData(unittest.TestCase):
             x.standard_name = 'projection_y_coordinate'
             y[:] = np.arange(1)
 
-            mtg_geos_projection = nc_byte.createVariable('mtg_geos_projection', np.int, dimensions=())
+            mtg_geos_projection = nc_byte.createVariable('mtg_geos_projection', int, dimensions=())
             mtg_geos_projection.longitude_of_projection_origin = 10.0
             mtg_geos_projection.semi_major_axis = 6378137.
             mtg_geos_projection.semi_minor_axis = 6356752.
